# Optimize documentation comment querying