@Pilchie 2016-10-01
tag @jasonmalinowski, @mattwar, @dotnet/roslyn-ide for review